Push ingest split on planning and executing
