+++ src/adapter/http/tests/tests/test_data_ingest.rs
@@ -603,6 +603,7 @@ impl DataIngestHarness {
         let catalog = dill::CatalogBuilder::new()
             .add::<DataFormatRegistryImpl>()
             .add::<PushIngestServiceImpl>()
+            .add::<PushIngestPlannerImpl>()
             .add::<EngineProvisionerNull>()
             .add::<UploadServiceLocal>()
             .add_value(FileUploadLimitConfig::new_in_bytes(1000))
