+++ src/domain/core/src/services/ingest/mod.rs
@@ -11,6 +11,7 @@ mod data_format_registry;
 mod data_writer;
 mod merge_strategy;
 mod polling_ingest_service;
+mod push_ingest_planner;
 mod push_ingest_service;
 mod reader;
 
@@ -18,5 +19,6 @@ pub use data_format_registry::*;
 pub use data_writer::*;
 pub use merge_strategy::*;
 pub use polling_ingest_service::*;
+pub use push_ingest_planner::*;
 pub use push_ingest_service::*;
 pub use reader::*;
