+++ src/infra/core/src/ingest/push_ingest_service_impl.rs
@@ -16,9 +16,6 @@ use internal_error::{ErrorIntoInternal, InternalError, ResultIntoInternal};
 use kamu_core::ingest::*;
 use kamu_core::*;
 use kamu_ingest_datafusion::*;
-use opendatafabric::*;
-use random_names::get_random_name;
-use time_source::SystemTimeSource;
 use tokio::io::AsyncRead;
 
 use super::ingest_common;
@@ -28,9 +25,7 @@ use super::ingest_common;
 pub struct PushIngestServiceImpl {
     object_store_registry: Arc<dyn ObjectStoreRegistry>,
     data_format_registry: Arc<dyn DataFormatRegistry>,
-    time_source: Arc<dyn SystemTimeSource>,
     engine_provisioner: Arc<dyn EngineProvisioner>,
-    run_info_dir: Arc<RunInfoDir>,
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -41,75 +36,12 @@ impl PushIngestServiceImpl {
     pub fn new(
         object_store_registry: Arc<dyn ObjectStoreRegistry>,
         data_format_registry: Arc<dyn DataFormatRegistry>,
-        time_source: Arc<dyn SystemTimeSource>,
         engine_provisioner: Arc<dyn EngineProvisioner>,
-        run_info_dir: Arc<RunInfoDir>,
     ) -> Self {
         Self {
             object_store_registry,
             data_format_registry,
-            time_source,
             engine_provisioner,
-            run_info_dir,
-        }
-    }
-
-    async fn prepare_metadata_state(
-        &self,
-        target: ResolvedDataset,
-        source_name: Option<&str>,
-    ) -> Result<DataWriterMetadataState, PushIngestPlanningError> {
-        let metadata_state = DataWriterMetadataState::build(target, &BlockRef::Head, source_name)
-            .await
-            .map_err(|e| match e {
-                ScanMetadataError::SourceNotFound(err) => {
-                    PushIngestPlanningError::SourceNotFound(err.into())
-                }
-                ScanMetadataError::Internal(err) => PushIngestPlanningError::Internal(err),
-            })?;
-        Ok(metadata_state)
-    }
-
-    async fn auto_create_push_source(
-        &self,
-        target: ResolvedDataset,
-        source_name: &str,
-        opts: &PushIngestOpts,
-    ) -> Result<AddPushSource, PushIngestPlanningError> {
-        let read = match &opts.media_type {
-            Some(media_type) => {
-                match self
-                    .data_format_registry
-                    .get_best_effort_config(None, media_type)
-                {
-                    Ok(read_step) => Ok(read_step),
-                    Err(e) => Err(PushIngestPlanningError::UnsupportedMediaType(e)),
-                }
-            }
-            None => Err(PushIngestPlanningError::SourceNotFound(
-                PushSourceNotFoundError::new(Some(source_name)),
-            )),
-        }?;
-
-        let add_push_source_event = AddPushSource {
-            source_name: String::from("auto"),
-            read,
-            preprocess: None,
-            merge: opendatafabric::MergeStrategy::Append(opendatafabric::MergeStrategyAppend {}),
-        };
-
-        let commit_result = target
-            .commit_event(
-                MetadataEvent::AddPushSource(add_push_source_event.clone()),
-                CommitOpts {
-                    system_time: opts.source_event_time,
-                    ..CommitOpts::default()
-                },
-            )
-            .await;
-        match commit_result {
-            Ok(_) => Ok(add_push_source_event),
-            Err(e) => Err(PushIngestPlanningError::CommitError(e)),
         }
     }
 
@@ -376,54 +308,6 @@ impl PushIngestServiceImpl {
 
 #[async_trait::async_trait]
 impl PushIngestService for PushIngestServiceImpl {
-    /// Uses or auto-creates push source definition in metadata to plan
-    /// ingestion
-    async fn plan_ingest(
-        &self,
-        target: ResolvedDataset,
-        source_name: Option<&str>,
-        opts: PushIngestOpts,
-    ) -> Result<PushIngestPlan, PushIngestPlanningError> {
-        let mut metadata_state = self
-            .prepare_metadata_state(target.clone(), source_name)
-            .await?;
-
-        let push_source = match (&metadata_state.source_event, opts.auto_create_push_source) {
-            // No push source, and it's allowed to create
-            (None, true) => {
-                let add_push_source_event = self
-                    .auto_create_push_source(target.clone(), "auto", &opts)
-                    .await?;
-
-                // Update data writer, as we've modified the dataset
-                metadata_state = self.prepare_metadata_state(target, source_name).await?;
-                Ok(add_push_source_event)
-            }
-
-            // Got existing push source
-            (Some(MetadataEvent::AddPushSource(e)), _) => Ok(e.clone()),
-
-            // No push source and not allowed to create
-            _ => Err(PushIngestPlanningError::SourceNotFound(
-                PushSourceNotFoundError::new(source_name),
-            )),
-        }?;
-
-        let operation_id = get_random_name(None, 10);
-        let operation_dir = self.run_info_dir.join(format!("ingest-{operation_id}"));
-
-        Ok(PushIngestPlan {
-            args: PushIngestArgs {
-                operation_id,
-                operation_dir,
-                system_time: self.time_source.now(),
-                opts,
-                push_source,
-            },
-            metadata_state: Box::new(metadata_state),
-        })
-    }
-
     #[tracing::instrument(level = "info", skip_all, fields(target=%target.get_handle()))]
     async fn ingest_from_url(
         &self,
